+++ R/get_circumscribingcircle.R
@@ -3,7 +3,9 @@
 #' @description Calculates the diameter of the smallest circumscribing circle around patches in a landscape.
 #'
 #' @param landscape RasterLayer or matrix (with x,y,id columns)
-#' @param resolution Resolution of the landscape (only needed if matrix as input is used)
+#' @param resolution_x Resolution of the landscape (only needed if matrix as input is used)
+#' @param resolution_y Resolution of the landscape (only needed if matrix as input is used)
+
 #'
 #' @details
 #' Fast and memory safe Rcpp implementation for calculating maximum euclidean distances between
@@ -19,29 +21,36 @@
 #'
 #' @examples
 #' # get patches for class 1 from testdata as raster
-#' class_1 <- get_patches(landscape,1)[[1]]
+#' class_1 <- get_patches(landscape, class = 1)[[1]]
 #'
 #' # calculate the max distance between cell edges of each class
 #' get_circumscribingcircle(class_1)
 #'
-#' # do the same with a 3 column matrix (x,y,id)
+#' # do the same with a 3 column matrix (x, y, id)
 #' class_1_matrix <- raster::rasterToPoints(class_1)
-#' get_circumscribingcircle(class_1_matrix, 1)
+#' get_circumscribingcircle(class_1_matrix, resolution_x = 1, resolution_y = 1)
 #'
 #' @aliases get_circumscribingcircle
 #' @rdname get_circumscribingcircle
 #'
 #' @export
-get_circumscribingcircle <- function(landscape, resolution) UseMethod("get_circumscribingcircle")
+get_circumscribingcircle <- function(landscape, resolution_x, resolution_y) UseMethod("get_circumscribingcircle")
 
 #' @name get_circumscribingcircle
 #' @export
-get_circumscribingcircle.RasterLayer <- function(landscape, resolution) {
+get_circumscribingcircle.RasterLayer <- function(landscape,
+                                                 resolution_x = NULL,
+                                                 resolution_y = NULL) {
+
     points_mat <- raster::rasterToPoints(landscape)
 
-    resol <- raster::res(landscape)
+    resolution_xy <- raster::res(landscape)
+    resolution_x <- resolution_xy[[1]]
+    resolution_y <- resolution_xy[[2]]
 
-    circle <- rcpp_get_circle(points_mat, resolution = prod(resol))
+    circle <- rcpp_get_circle(points_mat,
+                              resolution_x = resolution_x,
+                              resolution_y = resolution_y)
 
     tibble::tibble(id = circle[, 1],
                    dist = circle[, 2])
@@ -50,17 +59,21 @@ get_circumscribingcircle.RasterLayer <- function(landscape, resolution) {
 
 #' @name get_circumscribingcircle
 #' @export
-get_circumscribingcircle.matrix <- function(landscape, resolution = NULL) {
+get_circumscribingcircle.matrix <- function(landscape,
+                                            resolution_x = NULL,
+                                            resolution_y = NULL) {
 
     if ( ncol(landscape) != 3){
-        stop("Coordinate matrix must have 3 (x,y,id) columns.", call. = TRUE)
+        stop("Coordinate matrix must have 3 (x, y, id) columns.", call. = TRUE)
     }
 
-    if (is.null(resolution)){
+    if (is.null(resolution_x) || is.null(resolution_y)){
         stop("Resolution must be provided to correctly calculate the edges. ", call. = TRUE)
     }
 
-    circle <- rcpp_get_circle(landscape, resolution = resolution)
+    circle <- rcpp_get_circle(landscape,
+                              resolution_x = resolution_x,
+                              resolution_y = resolution_y)
 
     tibble::tibble(id = circle[, 1],
                    dist = circle[, 2])