+++ R/lsm_p_core.R
@@ -134,21 +134,35 @@ lsm_p_core.list <- function(landscape, directions = 8,
 
 lsm_p_core_calc <- function(landscape, directions, consider_boundary, edge_depth) {
 
-    landscape_labeled <- get_patches(landscape, directions = directions)
+    classes <- rcpp_get_unique_values(raster::as.matrix(landscape))
 
-    core <- lapply(landscape_labeled, function(patches_class) {
+    resolution_xy <- raster::res(landscape)
+    landscape_padded_extent <- raster::extent(landscape) + (2 * resolution_xy)
+    landscape_labeled_empty <- raster::raster(x = landscape_padded_extent,
+                                              resolution = resolution_xy,
+                                              crs = raster::crs(landscape))
+
+    core <- lapply(classes, function(patches_class) {
+
+        landscape_labeled <- get_patches(landscape,
+                                         class = patches_class,
+                                         directions = directions)[[1]]
 
         if(!isTRUE(consider_boundary)) {
-            patches_class <- pad_raster(patches_class, pad_raster_value = NA,
-                                        pad_raster_cells = 1,
-                                        global = FALSE)
+
+            landscape_padded <- pad_raster(landscape_labeled,
+                                           pad_raster_value = NA,
+                                           pad_raster_cells = 1,
+                                           global = FALSE)
+
+            landscape_labeled <- raster::setValues(landscape_labeled_empty, landscape_padded)
         }
 
-        class_edge <- raster::boundaries(patches_class,
-                                       directions = 4)
+        class_edge <- raster::boundaries(landscape_labeled,
+                                         directions = 4)
 
-        cells_edge_patch <- table(factor(raster::values(patches_class)[raster::values(class_edge) == 1],
-                                   levels = unique(raster::values(patches_class))))
+        cells_edge_patch <- table(factor(raster::values(landscape_labeled)[raster::values(class_edge) == 1],
+                                  levels = rcpp_get_unique_values(raster::as.matrix(landscape_labeled))))
 
         if(edge_depth > 1){
             for(i in seq_len(edge_depth - 1)){
@@ -156,21 +170,19 @@ lsm_p_core_calc <- function(landscape, directions, consider_boundary, edge_depth
                 raster::values(class_edge)[raster::values(class_edge) == 1] <- NA
 
                 class_edge <- raster::boundaries(class_edge,
-                                               directions = 4)
+                                                 directions = 4)
 
-                cells_edge_patch <- cells_edge_patch + table(factor(raster::values(patches_class)[raster::values(class_edge) == 1],
-                                                              levels = unique(raster::values(patches_class))))
+                cells_edge_patch <- cells_edge_patch + table(factor(raster::values(landscape_labeled)[raster::values(class_edge) == 1],
+                                                             levels = rcpp_get_unique_values(raster::as.matrix(landscape_labeled))))
             }
         }
 
-        cells_patch <- table(factor(raster::values(patches_class),
-                                    levels = unique(raster::values(patches_class))))
+        cells_patch <- table(factor(raster::values(landscape_labeled),
+                                    levels = rcpp_get_unique_values(raster::as.matrix(landscape_labeled))))
 
-        core_area <- (cells_patch - cells_edge_patch) * prod(raster::res(patches_class)) / 10000
+        core_area <- (cells_patch - cells_edge_patch) * prod(resolution_xy) / 10000
 
-        class <- sub("Class_", "", names(patches_class))
-
-        tibble::tibble(class = class,
+        tibble::tibble(class = patches_class,
                        value = core_area)
     })