+++ tests/testthat/test-get-circumscribungcircle.R
@@ -1,20 +1,30 @@
 context("get_circumscribingcircle")
 
+class_1 <- get_patches(landscape, class = 1)[[1]]
+class_1_irr <- get_patches(landscape_diff_res, class = 1)[[1]]
 
-# get patches for class 1 from testdata as raster
-class_1 <- get_patches(landscape,1)[[1]]
-# calculate the distance between patches
-cc_rast <- get_circumscribingcircle(class_1)
-# do the same with a 3 column matrix (x,y,id)
-class_1_matrix <- raster::rasterToPoints(class_1)
-cc_mat<- get_circumscribingcircle(class_1_matrix, resolution = 1)
+test_that("get_circumscribingcircle has one radius for each patch", {
 
-test_that("get_adjacencies runs and returns a matrix", {
-    expect_is(cc_rast, "tbl_df")
-    expect_is(cc_mat, "tbl_df")
+    circle <- get_circumscribingcircle(class_1)
+    expect_equal(nrow(circle),
+                     expected =  max(get_unique_values(class_1)[[1]]))
+})
 
-    expect_true(round(cc_rast$dist[[1]],2) == 6.28)
-    expect_true(round(cc_mat$dist[[1]],2) == 6.28)
+test_that("get_circumscribingcircle has two rows and is tibble", {
 
-    expect_error(get_circumscribingcircle(class_1_matrix))
+    circle <- get_circumscribingcircle(class_1)
+    expect_is(circle, class = "tbl")
+    expect_equal(ncol(circle), expected = 2)
+
+})
+
+test_that("get_circumscribingcircle works for irregular raster", {
+
+    circle <- get_circumscribingcircle(class_1_irr)
+
+    expect_is(circle, class = "tbl")
+    expect_equal(ncol(circle), expected = 2)
+    expect_equal(nrow(circle),
+                 expected =  max(get_unique_values(class_1_irr)[[1]]))
+
 })